Refactor theme reloading mechanism to support extension updates and add reload annotation
