+++ application/src/main/java/run/halo/app/theme/service/ThemeServiceImpl.java
@@ -11,11 +11,10 @@ import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.time.Duration;
+import java.time.Instant;
 import java.util.HashMap;
-import java.util.HashSet;
 import java.util.Map;
 import java.util.Objects;
-import java.util.Set;
 import java.util.function.Predicate;
 import lombok.AllArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -34,7 +33,6 @@ import org.springframework.util.ResourceUtils;
 import org.springframework.util.StreamUtils;
 import org.springframework.web.server.ServerErrorException;
 import org.springframework.web.server.ServerWebInputException;
-import reactor.core.publisher.Flux;
 import reactor.core.publisher.Mono;
 import reactor.core.scheduler.Scheduler;
 import reactor.core.scheduler.Schedulers;
@@ -42,10 +40,7 @@ import reactor.util.retry.Retry;
 import run.halo.app.core.extension.AnnotationSetting;
 import run.halo.app.core.extension.Setting;
 import run.halo.app.core.extension.Theme;
-import run.halo.app.core.extension.notification.NotificationTemplate;
 import run.halo.app.extension.ConfigMap;
-import run.halo.app.extension.Extension;
-import run.halo.app.extension.GroupVersionKind;
 import run.halo.app.extension.MetadataUtil;
 import run.halo.app.extension.ReactiveExtensionClient;
 import run.halo.app.extension.Unstructured;
@@ -205,7 +200,7 @@ public class ThemeServiceImpl implements ThemeService {
                         new String[] {requires, systemVersion});
                 }
             })
-            .flatMap(theme -> {
+            .delayUntil(theme -> {
                 var unstructureds = ThemeUtils.loadThemeResources(getThemePath(theme));
                 if (unstructureds.stream()
                     .filter(hasSettingsYaml(theme))
@@ -219,18 +214,7 @@ public class ThemeServiceImpl implements ThemeService {
                     return Mono.error(new IllegalStateException(
                         "Theme must only have one config.yaml or config.yml."));
                 }
-                return Flux.fromIterable(unstructureds)
-                    .filter(unstructured -> ExtensionWhitelist.of(theme).isAllowed(unstructured))
-                    .doOnNext(unstructured ->
-                        populateThemeNameLabel(unstructured, theme.getMetadata().getName()))
-                    .flatMap(unstructured -> {
-                        if (isUpgrade) {
-                            return createOrUpdate(unstructured);
-                        }
-                        return client.create(unstructured);
-                    })
-                    .then()
-                    .thenReturn(theme);
+                return Mono.empty();
             });
     }
 
@@ -251,43 +235,24 @@ public class ThemeServiceImpl implements ThemeService {
     @Override
     public Mono<Theme> reloadTheme(String name) {
         return client.fetch(Theme.class, name)
-            .flatMap(oldTheme -> {
+            .delayUntil(oldTheme -> {
                 String settingName = oldTheme.getSpec().getSettingName();
                 return waitForSettingDeleted(settingName)
                     .then(waitForAnnotationSettingsDeleted(name));
             })
-            .then(Mono.defer(() -> {
-                Path themePath = themeRoot.get().resolve(name);
-                Path themeManifestPath = ThemeUtils.resolveThemeManifest(themePath);
+            .flatMap(oldTheme -> {
+                var themePath = themeRoot.get().resolve(name);
+                var themeManifestPath = ThemeUtils.resolveThemeManifest(themePath);
                 if (themeManifestPath == null) {
-                    throw new IllegalArgumentException(
-                        "The manifest file [theme.yaml] is required.");
+                    return Mono.error(new ServerWebInputException(
+                        "The manifest file [theme.yaml] is required."
+                    ));
                 }
-                Unstructured unstructured = loadThemeManifest(themeManifestPath);
-                Theme newTheme = Unstructured.OBJECT_MAPPER.convertValue(unstructured,
-                    Theme.class);
-                return client.fetch(Theme.class, name)
-                    .map(oldTheme -> {
-                        newTheme.getMetadata().setVersion(oldTheme.getMetadata().getVersion());
-                        return newTheme;
-                    })
-                    .flatMap(client::update);
-            }))
-            .flatMap(theme -> Flux.fromIterable(ThemeUtils.loadThemeResources(getThemePath(theme)))
-                .filter(unstructured -> ExtensionWhitelist.of(theme).isAllowed(unstructured))
-                .doOnNext(unstructured -> populateThemeNameLabel(unstructured, name))
-                .flatMap(this::createOrUpdate)
-                .then(Mono.just(theme))
-            );
-    }
-
-    private static void populateThemeNameLabel(Unstructured unstructured, String themeName) {
-        Map<String, String> labels = unstructured.getMetadata().getLabels();
-        if (labels == null) {
-            labels = new HashMap<>();
-            unstructured.getMetadata().setLabels(labels);
-        }
-        labels.put(Theme.THEME_NAME_LABEL, themeName);
+                var unstructured = loadThemeManifest(themeManifestPath);
+                var newTheme = Unstructured.OBJECT_MAPPER.convertValue(unstructured, Theme.class);
+                updateTheme(oldTheme, newTheme);
+                return client.update(oldTheme);
+            });
     }
 
     @Override
@@ -392,63 +357,6 @@ public class ThemeServiceImpl implements ThemeService {
             .then();
     }
 
-    static class ExtensionWhitelist {
-        private final Set<AllowedExtension> rules;
-
-        private ExtensionWhitelist(Theme theme) {
-            this.rules = getRules(theme);
-        }
-
-        public static ExtensionWhitelist of(Theme theme) {
-            return new ExtensionWhitelist(theme);
-        }
-
-        public boolean isAllowed(Unstructured unstructured) {
-            return this.rules.stream()
-                .anyMatch(rule -> rule.matches(unstructured));
-        }
-
-        private Set<AllowedExtension> getRules(Theme theme) {
-            var rules = new HashSet<AllowedExtension>();
-            rules.add(AllowedExtension.of(AnnotationSetting.class));
-            rules.add(AllowedExtension.of(NotificationTemplate.class));
-
-            var configMapName = theme.getSpec().getConfigMapName();
-            if (StringUtils.isNotBlank(configMapName)) {
-                rules.add(AllowedExtension.of(ConfigMap.class, configMapName));
-            }
-
-            var settingName = theme.getSpec().getSettingName();
-            if (StringUtils.isNotBlank(settingName)) {
-                rules.add(AllowedExtension.of(Setting.class, settingName));
-            }
-            return rules;
-        }
-    }
-
-    record AllowedExtension(String apiGroup, String kind, String name) {
-        AllowedExtension {
-            Assert.notNull(apiGroup, "The apiGroup must not be null");
-            Assert.notNull(kind, "Kind must not be null");
-        }
-
-        public static <E extends Extension> AllowedExtension of(Class<E> clazz) {
-            return of(clazz, null);
-        }
-
-        public static <E extends Extension> AllowedExtension of(Class<E> clazz, String name) {
-            var gvk = GroupVersionKind.fromExtension(clazz);
-            return new AllowedExtension(gvk.group(), gvk.kind(), name);
-        }
-
-        public boolean matches(Unstructured unstructured) {
-            var groupVersionKind = unstructured.groupVersionKind();
-            return this.apiGroup.equals(groupVersionKind.group())
-                && this.kind.equals(groupVersionKind.kind())
-                && (this.name == null || this.name.equals(unstructured.getMetadata().getName()));
-        }
-    }
-
     private static void updateTheme(Theme existing, Theme updating) {
         var existingSpec = existing.getSpec();
         var updatingSpec = updating.getSpec();
@@ -476,11 +384,15 @@ public class ThemeServiceImpl implements ThemeService {
             }
             existingMeta.getLabels().putAll(updatingMeta.getLabels());
         }
+        if (existingMeta.getAnnotations() == null) {
+            existingMeta.setAnnotations(new HashMap<>());
+        }
+        // make sure the theme will be reloaded after upgrade by adding a special annotation with
+        // current timestamp
+        existingMeta.getAnnotations()
+            .put(Theme.REQUEST_RELOAD_ANNOTATION, Instant.now().toString());
         // merge annotations
         if (updatingMeta.getAnnotations() != null) {
-            if (existingMeta.getAnnotations() == null) {
-                existingMeta.setAnnotations(new HashMap<>());
-            }
             existingMeta.getAnnotations().putAll(updatingMeta.getAnnotations());
         }
     }
